requirements.txt
streamlit
tensorflow
opencv-python
matplotlib
numpy
Pillow